Transfer the type of city to const
