+++ src/const.js
@@ -1,4 +1,65 @@
-export const Offers = {
+const Cities = [
+  `Amsterdam`,
+  `Geneva`,
+  `Barcelona`,
+  `Dresden`,
+  `Praha`
+];
+
+const Types = [
+  {
+    id: `taxi`,
+    title: `Taxi`,
+    placeholder: `to`
+  },
+  {
+    id: `bus`,
+    title: `Bus`,
+    placeholder: `to`
+  },
+  {
+    id: `train`,
+    title: `Train`,
+    placeholder: `to`
+  },
+  {
+    id: `ship`,
+    title: `Ship`,
+    placeholder: `to`
+  },
+  {
+    id: `transport`,
+    title: `Transport`,
+    placeholder: `to`
+  },
+  {
+    id: `drive`,
+    title: `Drive`,
+    placeholder: `to`
+  },
+  {
+    id: `flight`,
+    title: `Flight`,
+    placeholder: `to`
+  },
+  {
+    id: `check-in`,
+    title: `Check`,
+    placeholder: `into`
+  },
+  {
+    id: `sightseeing`,
+    title: `Sightseeing`,
+    placeholder: `at`
+  },
+  {
+    id: `restaurant`,
+    title: `Restaurant`,
+    placeholder: `at`
+  },
+];
+
+const Offers = {
   luggage:
   {
     title: `Add luggage`,
@@ -24,4 +85,10 @@ export const Offers = {
     title: `Travel by train`,
     price: 40,
   }
+};
+
+export {
+  Cities,
+  Types,
+  Offers
 };
